+++ cloudwatch-doom.el
@@ -5,8 +5,8 @@
 ;; Author: Randol Reeves <randol.reeves+emacs@gmail.com>
 ;; Maintainer: Randol Reeves <randol.reeves+emacs@gmail.com>
 ;; Created: November 04, 2025
-;; Modified: November 04, 2025
-;; Version: 0.2.1
+;; Modified: January 29, 2026
+;; Version: 0.4.0
 ;; Keywords: tools aws cloudwatch logs monitoring devops kubernetes observability
 ;; Homepage: https://github.com/rand-fu/cloudwatch-el
 ;; Package-Requires: ((emacs "27.1") (cloudwatch "0.1.0"))